@@ -2,7 +2,7 @@ export 'user_data.dart';
 
 class UserData {
 
-static String token = "token:bhjbtyBHgtyvytyv";
+static const token = "token:bhjbtyBHgtyvytyv";
 
 }
 
