@@ -1,33 +1,58 @@
-import 'package:meta/meta.dart';
-import 'package:flutter_network/movie/movie.dart';
-import '../models/movie.dart';
+import 'package:sec_2/movie/index.dart';
+
+import '../../utils/either.dart';
 
 class MovieRepository {
-  final MovieDataProvider dataProvider;
+  MovieProvider movieProvider;
+  MovieRepository(this.movieProvider);
 
-  MovieRepository({@required this.dataProvider}) : assert(dataProvider != null);
+  // Future<List<Movie>> getMoviesWithPriority(int priority) async {
+  //   final movies = await movieProvider.getAllMovies();
+  //   return movies.where((movie) => movie.priority == priority).toList();
+  // }
 
-  Future<Movie> createMovie(Movie movie) async {
-    return await dataProvider.createMovie(movie);
+  Future<Either<List<Movie>>> getAllMovies() async {
+    try {
+      final movies = await movieProvider.getAllMovies();
+      return Either(val: movies);
+    } catch (err) {
+      return Either(error: "Couldn't load movies");
+    }
   }
 
-  Future<List<Movie>> getMovies() async {
-    //remove
-    // print("movie repo lounched");
-    // List<Movie> test = dataProvider.getMovies();
-    // print(dataProvider.getMovies());
-    //remove
-    // return await dataProvider.getMovies();
-    List<Movie> c = await dataProvider.getMovies();
-    print(0);
-    return c;
-  }
+  // Future<Either<Movie>> getMovie(int id) async {
+  //   try {
+  //     final movie = await movieProvider.getMovie(id);
+  //     return Either(val: movie);
+  //   } catch (err) {
+  //     return Either(error: "Movie not found");
+  //   }
+  // }
 
-  Future<void> updateMovie(Movie movie) async {
-    await dataProvider.updateMovie(movie);
-  }
+  // addMovie(Movie movie) async {
+  //   // check for profanity
+  //   // movie.description.contains("sidib");
+  //   final newMovie = Movie(
+  //     id: movie.id,
+  //     priority: movie.priority,
+  //     description: movie.description.trim(),
+  //   );
 
-  Future<void> deleteMovie(String id) async {
-    await dataProvider.deleteMovie(id);
-  }
+  //   await movieProvider.addMovie(newMovie);
+  // }
+
+  // Future<Either<String>> editMovie(String id, Movie movie) async {
+  //   try {
+  //     final newMovie = Movie(
+  //       id: movie.id,
+  //       priority: movie.priority,
+  //       description: movie.description.trim(),
+  //     );
+
+  //     await movieProvider.editMovie(id, newMovie);
+  //     return Either(val: "");
+  //   } catch (err) {
+  //     return Either(error: "Movie not found");
+  //   }
+  // }
 }
